Prepares to implement domain updates
