+++ common/isolationgroup/defaultisolationgroup/state_test.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package isolationgroup
+package defaultisolationgroup
 
 import (
 	"context"
@@ -209,106 +209,6 @@ func TestAvailableIsolationGroups(t *testing.T) {
 	for name, td := range tests {
 		t.Run(name, func(t *testing.T) {
 			assert.Equal(t, td.expected, availableIG(all, td.globalIGCfg, td.domainIGCfg))
-		})
-	}
-}
-
-func TestUpdateGlobalState(t *testing.T) {
-
-	tests := map[string]struct {
-		in           types.UpdateGlobalIsolationGroupsRequest
-		dcAffordance func(client *dynamicconfig.MockClient)
-		expectedErr  error
-	}{
-		"updating normal value": {
-			in: types.UpdateGlobalIsolationGroupsRequest{IsolationGroups: types.IsolationGroupConfiguration{
-				"zone-1": {Name: "zone-1", State: types.IsolationGroupStateHealthy},
-				"zone-2": {Name: "zone-2", State: types.IsolationGroupStateDrained},
-			}},
-			dcAffordance: func(client *dynamicconfig.MockClient) {
-				client.EXPECT().UpdateValue(
-					dynamicconfig.DefaultIsolationGroupConfigStoreManagerGlobalMapping,
-					gomock.Any(), // covering the mapping in the mapper unit-test instead
-				)
-			},
-		},
-	}
-
-	for name, td := range tests {
-		t.Run(name, func(t *testing.T) {
-			dcMock := dynamicconfig.NewMockClient(gomock.NewController(t))
-			td.dcAffordance(dcMock)
-			handler := defaultIsolationGroupStateHandler{
-				log:                        loggerimpl.NewNopLogger(),
-				globalIsolationGroupDrains: dcMock,
-			}
-			err := handler.UpdateGlobalState(context.Background(), td.in)
-			assert.Equal(t, td.expectedErr, err)
-		})
-	}
-}
-
-func TestGetGlobalState(t *testing.T) {
-
-	validInput := types.IsolationGroupConfiguration{
-		"zone-1": types.IsolationGroupPartition{
-			Name:  "zone-1",
-			State: types.IsolationGroupStateDrained,
-		},
-		"zone-2": types.IsolationGroupPartition{
-			Name:  "zone-2",
-			State: types.IsolationGroupStateHealthy,
-		},
-	}
-
-	validCfg, _ := mapUpdateGlobalIsolationGroupsRequest(validInput)
-
-	validCfgData := validCfg[0].Value.GetData()
-	dynamicConfigResponse := []interface{}{}
-	json.Unmarshal(validCfgData, &dynamicConfigResponse)
-
-	tests := map[string]struct {
-		in           types.GetGlobalIsolationGroupsRequest
-		dcAffordance func(client *dynamicconfig.MockClient)
-		expected     *types.GetGlobalIsolationGroupsResponse
-		expectedErr  error
-	}{
-		"updating normal value": {
-			in: types.GetGlobalIsolationGroupsRequest{},
-			dcAffordance: func(client *dynamicconfig.MockClient) {
-				client.EXPECT().GetListValue(
-					dynamicconfig.DefaultIsolationGroupConfigStoreManagerGlobalMapping,
-					gomock.Any(),
-				).Return(dynamicConfigResponse, nil)
-			},
-			expected: &types.GetGlobalIsolationGroupsResponse{IsolationGroups: validInput},
-		},
-		"an error was returned": {
-			in: types.GetGlobalIsolationGroupsRequest{},
-			dcAffordance: func(client *dynamicconfig.MockClient) {
-				client.EXPECT().GetListValue(
-					dynamicconfig.DefaultIsolationGroupConfigStoreManagerGlobalMapping,
-					gomock.Any(),
-				).Return(nil, errors.New("an error"))
-			},
-			expectedErr: errors.New("failed to get global isolation groups from datastore: an error"),
-		},
-	}
-
-	for name, td := range tests {
-		t.Run(name, func(t *testing.T) {
-			dcMock := dynamicconfig.NewMockClient(gomock.NewController(t))
-			td.dcAffordance(dcMock)
-			handler := defaultIsolationGroupStateHandler{
-				log:                        loggerimpl.NewNopLogger(),
-				globalIsolationGroupDrains: dcMock,
-			}
-			res, err := handler.GetGlobalState(context.Background())
-			assert.Equal(t, td.expected, res)
-			if td.expectedErr != nil {
-				// only compare strings because full wrapping makes it fiddly otherwise
-				assert.Equal(t, td.expectedErr.Error(), err.Error())
-			}
 		})
 	}
 }
